==> 00-Exploracion-archivos.ipynb <==
# Universidad del Desarrollo
## Magister Data Science - Macroeconomía
### Grupo N°2: (Región de los Lagos)
* Jose Pedro
* Cristian
* Paul
* Max
* Hernan

## Exploracion de archivos

**Objetivo**

Obtener la serie de precios de los productos agregados por sector


import pandas as pd

### 1. Se cargan los productos de la ODEPA

df_01 = pd.read_excel('input/aceite.xlsx')
df_02 = pd.read_excel('input/azucar.xlsx')
df_03 = pd.read_excel('input/carnes.xlsx')
df_04 = pd.read_excel('input/cereales.xlsx')
df_05 = pd.read_excel('input/frutas.xlsx')
df_06 = pd.read_excel('input/harina.xlsx')
df_07 = pd.read_excel('input/hortalizas.xlsx')
df_08 = pd.read_excel('input/huevos.xlsx')
df_09 = pd.read_excel('input/lacteos.xlsx')
df_10 = pd.read_excel('input/legumbres.xlsx')
df_11 = pd.read_excel('input/pan.xlsx')

df_01["Tipo"] = "Aceite"
df_02["Tipo"] = "Azucar"
df_03["Tipo"] = "Carnes"
df_04["Tipo"] = "Cereales"
df_05["Tipo"] = "Frutas"
df_06["Tipo"] = "Harina"
df_07["Tipo"] = "Hortaliza"
df_08["Tipo"] = "Huevos"
df_09["Tipo"] = "Lacteos"
df_10["Tipo"] = "Legumbre"
df_11["Tipo"] = "Pan"

El dataset de Azucar requiere un arreglo adicional, dado que no viene columna *Producto*

df_02['Producto'] = 'Azucar normal'

dfs = [df_01, df_02, df_03, df_04, df_05, df_06, df_07, df_08, df_09, df_10, df_11]

### 2. Funcion para normalizar periodo

def GetPeriodo(mes):
    tmp = mes.split("/") 
    return tmp[1] + "-" + tmp[0]

### 3. Creacion de Dataset que corresponde a nuestra canasta

for df in dfs:
    df["Momento"] = df["Mes/Año"].map(GetPeriodo)

df_01.head()

### 4. Concatenamos archivos

pd.concat([df_01, df_02, df_03, df_04, df_05, df_06, df_07, df_08, df_09, df_10, df_11]).to_csv("output/rlagos_canasta.csv", index=False)
