==> 02-Imputacion.ipynb <==
# Universidad del Desarrollo
## Magister Data Science - Macroeconomía
### Grupo N°2: (Región de los Lagos)
* Jose Pedro
* Cristian
* Paul
* Max
* Hernan


# Imputación de los datos

import pandas as pd
import seaborn as sns
from scipy.stats.mstats import gmean

df = pd.read_csv('output/rlagos_canasta.csv')

df.head()

df.shape

df.Tipo.value_counts()

df[df.Tipo == 'Legumbre'].Producto.value_counts()

df[df.Tipo == 'Hortaliza'].Producto.value_counts()

df[df.Tipo == 'Frutas'].Producto.value_counts()

### Paso 1: buscar todos los promedios geometricos

Creamos un dataset nuevo con todas las combinaciones Tipo / Sector

Combos = df.groupby(['Tipo','Producto','Sector']).count().reset_index()[['Tipo','Producto','Sector']]

Combos.head()

Una prueba para ver como calculamos el promedio geometrico de una combinación a mano, antes de crear una función:

gmean(df[(df.Tipo == 'Aceite') & (df.Producto == 'Maravilla') & (df.Sector == 'Osorno')]['Precio promedio'])

Creamos una función que calcula el promedio geometrico en base al dataframe completo, pero filtrado por el dataframe Tipo / Producto / Sector

def calc_gmean(row):
    Tipo = row['Tipo']
    Producto = row['Producto']
    Sector = row['Sector']
    a = gmean(df[(df.Tipo == Tipo) & (df.Producto == Producto) & (df.Sector == Sector)]['Precio promedio'])
    return a

Aplicamos la función a cada fila del dataframe con las combinaciones Tipo / Producto / Sector

Combos['PromGeo'] = Combos.apply(calc_gmean, axis=1)

Combos.sample(10)

### Paso 2: Rellenar lo faltante

Estrategia:

1. Crear un dataframe con todas las fechas (momentos) del periodo.
1. A partir de cada columna en el dataframe *Combos*, crear un dataframe de una combinación Tipo/Producto/Sector
1. Hacer un merge entre el dataset de los momentos y el dataset Tipo/Producto/Sector para ver qué son los *momentos* que no tienen valores
1. Llenar los NaN con el promedio geometrico (y otros datos faltantes dado que estas filas no existían antes)
1. Crear un nuevo dataframe donde vamos agregando los dataframes de Tipo/Producto/Sector a medida que se completen.

df_momentos = pd.DataFrame(df['Momento'])
df_momentos.drop_duplicates(subset='Momento', keep='first', inplace=True)
df_momentos.reset_index()
df_momentos.shape

df_final = pd.DataFrame([])

def fill_nans(row):
    global df_final
    Tipo = row['Tipo']
    Producto = row['Producto']
    Sector = row['Sector']
    PromGeo = row['PromGeo']
    # Creamos un dataset con la combinación
    data = df[(df.Tipo == Tipo) & (df.Producto == Producto) & (df.Sector == Sector)][['Momento','Precio promedio','Tipo','Producto','Sector']]
    # Hacemos merge con el dataset que contiene todos los momentos
    data = pd.merge(df_momentos, data, on='Momento', how='left')
    # Llenamos los NaN
    data['Precio promedio'].fillna(PromGeo, inplace=True)
    data['Tipo'].fillna(Tipo, inplace=True)
    data['Producto'].fillna(Producto, inplace=True)
    data['Sector'].fillna(Sector, inplace=True)
    # Agregamos la data al dataframe completo
    df_final = df_final.append(data)

Combos.apply(fill_nans, axis=1)
df_final.head()

df_final.shape

df_final.reset_index(inplace=True)

Test con los que sabemos que estaban missing:

df_final[(df_final.Tipo == 'Hortaliza') & (df_final.Producto == 'Repollo') & (df_final.Sector == 'Frutillar') & (df_final.Momento == '2014-08')]

dx = df_final.groupby(["Momento", "Sector"])['Producto'].size().unstack()

mycolor=['black', 'red', 'green', 'blue', 'cyan']

dx.plot(kind='bar', stacked=True, figsize=(20, 6), color=mycolor, rot=45)

df = df_final[["Momento", "Sector", "Tipo", "Producto", "Precio promedio"]]

df.rename(columns={'Producto': 'Variedad', 'Precio promedio': 'Promedio', 'Tipo': 'Producto'}, inplace=True)

df.to_csv('output/canasta_ipc.csv', index=False)
