==> 03-Calculo-Precios-Producto.ipynb <==
# Universidad del Desarrollo
## Magister Data Science - Macroeconomía
### Grupo N°2: (Región de los Lagos)
* Jose Pedro
* Cristian
* Paul
* Max
* Hernan

## Calculo del Micro Indice de productos

**Objetivo**

El objetivo de este notebook es el calculo de la serie de precios de producto segun el sigueinte algoritmo.



$ 
\begin{equation*} 
p_i^t = \prod_{j=1}^n (\frac{v_j^t}{v_j^{t-1}})^{1/n} 
\end{equation*} 
$

Donde:
* $ p_i^t $ = Precio del producto i en el momento t
* $ p_i^0 $ = Precio del producto 0 en el mes base **(2014/08=100)**
* $ w_i   $ = Ponderación del producto i en la canasta
* $ v_j^t $ = Precio de la variaedad j en el momento t
* N = Numero de productos
* n = Numero de variedades

Dados los siguientes supuestos:
* n = 2
* t = 2014/09

Entonces:

P = ([ p_var1(2014-09) / p_var1(2014-08) ]$**$(1/2)) $*$ ([ p_var2(2014-09) / p_var2(2014-08) ]$**$(1/2))

import warnings
warnings.simplefilter(action='ignore')

import pandas as pd
import numpy as np
import calendar
import datetime

%matplotlib inline

### 1. Se cargan el dataset de productos

df = pd.read_csv('output/canasta_ipc.csv')

df.head()

df.Producto.value_counts()

### 2.- Transformamos Monento a un tipo de dato fecha para iterar sobre el periodo

df["Momento"] = pd.to_datetime(df["Momento"])

df.head()

* Años base:

base = datetime.date(2014, 8, 1)

### 3. Creamos una funcion para restar un mes al periodo de estudio

def add_months(sourcedate, months):
    month = sourcedate.month - 1 + months
    year = sourcedate.year + month // 12
    month = month % 12 + 1
    day = min(sourcedate.day, calendar.monthrange(year, month)[1])
    return datetime.date(year, month, day)

### 4. Creamos función para calculo de la serie factores de producto

def GetSerieProducto(df_input, group):
    v = []
    for name, g in group: 
        
        factor = 1
        n = g.Promedio.count()

        if g.Momento.iloc[0].date() > base:
            for row_num, row in g.iterrows(): 

                Pro1 = row["Promedio"]
                Mom2 = add_months(row["Momento"], -1)

                Pro2 = df_input[(df_input.Sector==row["Sector"]) & (df_input.Momento == Mom2) & (df_input.Variedad == row["Variedad"])]["Promedio"].iloc[0]

                factor *= (Pro1/Pro2)**(1/n)

        name = name + (factor,)  
        v.append(name)
    
    return pd.DataFrame(v, columns=['Momento', 'Sector', 'Producto', 'Factor'])

### 5. Calculo de factores de productos

#### 5.1 Aceite

df_aceite = df[df.Producto=="Aceite"]
group = df_aceite.groupby(['Momento', 'Sector', 'Producto'])

dfs_aceite = GetSerieProducto(df_aceite, group)

dfs_aceite.sample(3)

#### 5.2 Harina

df_harina = df[df.Producto=="Harina"]
group = df_harina.groupby(['Momento', 'Sector', 'Producto'])

dfs_harina = GetSerieProducto(df_harina, group)

dfs_harina.sample(3)

#### 5.3. Cereales

df_cereales = df[df.Producto=="Cereales"]
group = df_cereales.groupby(['Momento', 'Sector', 'Producto'])

dfs_cereales = GetSerieProducto(df_cereales, group)

dfs_cereales.sample(3)

#### 5.4 Azucar

df_azucar = df[df.Producto=="Azucar"]

group = df_azucar.groupby(['Momento', 'Sector', 'Producto'])

dfs_azucar = GetSerieProducto(df_azucar, group)

dfs_azucar.sample(3)

#### 5.5 Carne

df_carnes = df[df.Producto=="Carnes"]

group = df_carnes.groupby(['Momento', 'Sector', 'Producto'])

dfs_carnes = GetSerieProducto(df_carnes, group)

dfs_carnes.sample(3)

#### 5.6 Hortaliza

df_hortaliza = df[df.Producto=="Hortaliza"]

group = df_hortaliza.groupby(['Momento', 'Sector', 'Producto'])

dfs_hortaliza = GetSerieProducto(df_hortaliza, group)

dfs_hortaliza.sample(3)

#### 5.7 Frutas

df_frutas = df[df.Producto=="Frutas"]

group = df_frutas.groupby(['Momento', 'Sector', 'Producto'])

dfs_frutas = GetSerieProducto(df_frutas, group)

dfs_frutas.sample(3)

#### 5.8 Lacteos

df_lacteos = df[df.Producto=="Lacteos"]

group = df_lacteos.groupby(['Momento', 'Sector', 'Producto'])

dfs_lacteos = GetSerieProducto(df_lacteos, group)

dfs_lacteos.sample(3)

#### 5.9 Pan

df_pan = df[df.Producto=="Pan"]

group = df_pan.groupby(['Momento', 'Sector', 'Producto'])

dfs_pan = GetSerieProducto(df_pan, group)

dfs_pan.sample(3)

#### 5.10 Huevos

df_huevos = df[df.Producto=="Huevos"]

group = df_huevos.groupby(['Momento', 'Sector', 'Producto'])

dfs_huevos = GetSerieProducto(df_huevos, group)

dfs_huevos.sample(3)

#### 5.11 Legumbre

df_legumbre = df[df.Producto=="Legumbre"]

group = df_legumbre.groupby(['Momento', 'Sector', 'Producto'])

dfs_legumbre = GetSerieProducto(df_legumbre, group)

dfs_legumbre.sample(3)

### 6. Creamos dataset de precios

pd.concat([dfs_legumbre, 
           dfs_huevos, 
           dfs_pan, 
           dfs_lacteos, 
           dfs_frutas, 
           dfs_hortaliza, 
           dfs_carnes, 
           dfs_azucar, 
           dfs_cereales, 
           dfs_harina, 
           dfs_aceite]).to_csv("output/precios_productos.csv", index=False)
