==> 04 - Ponderacion.ipynb <==
# Ponderación

import pandas as pd

df_ine = pd.read_excel('input/ipc_producto_historico_julio2018.xlsx', header=3)

df_ine.head(2)

df_ine = df_ine[['PONDERACIÓN', 'GLOSA']]
df_ine.columns = [['Ponderacion', 'Producto_INE']]

len(df_ine)

df_ine.loc[0:3,:]

df_ine['Producto'] = df_ine['Producto_INE'].replace({'ARROZ':'Cereales',
                                                     'HARINA':'Harina',
                                                     'CEREAL':'Cereales',
                                                     'PAN':'Pan',
                                                     'PASTAS':'Cereales',
                                                     'CARNE DE VACUNO':'Carnes',
                                                     'CARNE DE CERDO':'Carnes',
                                                     'CARNE DE PAVO':'Carnes',
                                                     'CARNE DE POLLO':'Carnes',
                                                     'CECINAS':'Carnes',
                                                     'HAMBURGUESA':'Carnes',
                                                     'LECHE EN POLVO':'Lacteos',
                                                     'LECHE LÍQUIDA':'Lacteos',
                                                     'LECHE CONSERVADA':'Lacteos',
                                                     'YOGHURT':'Lacteos',
                                                     'POSTRES LÁCTEOS':'Lacteos',
                                                     'QUESO':'Lacteos',
                                                     'HUEVOS':'Huevos',
                                                     'MANTEQUILLA':'Lacteos',
                                                     'MARGARINA':'Lacteos',
                                                     'ACEITE VEGETAL':'Aceite',
                                                     'MANZANA':'Frutas',
                                                     'NARANJA':'Frutas',
                                                     'PERA':'Frutas',
                                                     'PLÁTANO':'Frutas',
                                                     'FRUTAS DE ESTACIÓN':'Frutas',
                                                     'ACELGA Y ESPINACA':'Hortaliza',
                                                     'CEBOLLA Y CEBOLLÍN':'Hortaliza',
                                                     'LECHUGA':'Hortaliza',
                                                     'LIMÓN':'Frutas',
                                                     'PALTA':'Frutas',
                                                     'TOMATE':'Hortaliza',
                                                     'ZANAHORIA':'Hortaliza',
                                                     'ZAPALLO':'Hortaliza',
                                                     'ZAPALLO ITALIANO':'Hortaliza',
                                                     'VERDURAS DE ESTACIÓN':'Hortaliza',
                                                     'LEGUMBRES':'Legumbre',
                                                     'PAPA':'Hortaliza',
                                                     'AZÚCAR':'Azucar'})

productos = ['ARROZ','HARINA','CEREAL','PAN','PASTAS','CARNE DE VACUNO','CARNE DE CERDO','CARNE DE PAVO','CARNE DE POLLO','CECINAS','HAMBURGUESA','LECHE EN POLVO','LECHE LÍQUIDA',
            'LECHE CONSERVADA','YOGHURT','POSTRES LÁCTEOS','QUESO','HUEVOS','MANTEQUILLA','MARGARINA','ACEITE VEGETAL','MANZANA','NARANJA','PERA','PLÁTANO','FRUTAS DE ESTACIÓN','ACELGA Y ESPINACA',
            'CEBOLLA Y CEBOLLÍN','LECHUGA','LIMÓN','PALTA','TOMATE','ZANAHORIA','ZAPALLO','ZAPALLO ITALIANO','VERDURAS DE ESTACIÓN','LEGUMBRES','PAPA','AZÚCAR']

df_ine.head()

productos = ['ARROZ','HARINA','CEREAL','PAN','PASTAS','CARNE DE VACUNO','CARNE DE CERDO','CARNE DE PAVO','CARNE DE POLLO','CECINAS','HAMBURGUESA','LECHE EN POLVO','LECHE LÍQUIDA',
            'LECHE CONSERVADA','YOGHURT','POSTRES LÁCTEOS','QUESO','HUEVOS','MANTEQUILLA','MARGARINA','ACEITE VEGETAL','MANZANA','NARANJA','PERA','PLÁTANO','FRUTAS DE ESTACIÓN','ACELGA Y ESPINACA',
            'CEBOLLA Y CEBOLLÍN','LECHUGA','LIMÓN','PALTA','TOMATE','ZANAHORIA','ZAPALLO','ZAPALLO ITALIANO','VERDURAS DE ESTACIÓN','LEGUMBRES','PAPA','AZÚCAR']

def cleanup(row):
    global productos
    if row['Producto_INE'].isin(productos).iloc[0] == True:
        return 'Mantener'
    else:
        return 'Desechar'

df_ine['Accion'] = df_ine.apply(cleanup, axis=1)

df_ine.to_csv("output/ponderados.csv", index=False)

df = pd.read_csv("output/ponderados.csv")

df.head(5)

df.Accion.value_counts()

df = df[df.Accion=="Mantener"]

df.drop(['Producto_INE','Accion'], axis=1, inplace=True)

df.Producto.value_counts()

df = df.groupby(['Producto']).sum()

df.sum(axis=0)

df_ine.Ponderacion.sum(axis=0)

df

suma_antes = df.sum(axis=0)

df['Ponderacion_Nueva'] =  df['Ponderacion'] / 13.35292 * 100

df

df.sum(axis=0)

df.reset_index(inplace=True)

df.to_csv("output/w_ponderados.csv", index=False)
