==> 04-Calculo-Indice-Precios-Consumidor - RL.ipynb <==
# Universidad del Desarrollo
## Magister Data Science - Macroeconomía
### Grupo N°2: (Región de los Lagos)
* Jose Pedro
* Cristian
* Paul
* Max
* Hernan

## Calculo Indice Precios al Consumidor

**Objetivo**

El objetivo de este notebook es el calculo de IPC según el siguiente algortimo.



$ 
\begin{equation*}
IPC_t   = |\sum_{k=1}^N \frac{p_i^t}{p_i^0} w_i| * 100
\end{equation*} 
$

Donde:
* $ p_i^t $ = Precio del producto i en el momento t
* $ p_i^0 $ = Precio del producto 0 en el mes base **(2014/08=100)**
* $ w_i   $ = Ponderación del producto i en la canasta
* N = Numero de productos

import pandas as pd
import numpy as np
import calendar
import datetime

%matplotlib inline

base = "2014-08"

### 1. Se cargan el dataset de productos

df = pd.read_csv('output/precios_productos.csv')

df.sample(10)

df.Sector.value_counts()

df_w = pd.read_csv("output/w_ponderados.csv")

df_w.head(11)

### 2. Calculamos precios base

dfp = pd.read_csv("output/canasta_ipc.csv")

dfp.head()

df_pb = dfp[dfp.Momento==base]

df_pb.sample(5)

pb = df_pb.groupby(["Producto"]).mean()
pb

### 3. IPC Osorno

df_osorno = df[(df.Sector=="Osorno") & (df.Momento > base + "-01")]

df_osorno.Sector.value_counts()

def GetIPC(df, df_base, df_p):
    v = []
    for o in df.Momento:
        ipc = 1
        prod = df[df.Momento == o].Producto
        for row in prod:
            p_0 = df_base.loc[row].iloc[0]
            p_j = df[(df.Momento == o) & (df.Producto == row)].Factor.iloc[0]

            w_i = df_p[df_p.Producto == row]["Ponderacion_Nueva"].iloc[0]
            
            ipc =+ ((p_j/p_0)*w_i)

        v.append({"ipc": ipc*100, "Momento": o})
        
    return pd.DataFrame(v)

df_ipc = GetIPC(df_osorno, pb, df_w)

df_ipc.head()

#df_ipc.plot(kind="bar")
